+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 Our first stable release. The package has been in beta for a few months without any major issues. We will keep on improving the codebase and continue to add more sync functionality.
 
 - Added the ability for the `smtc_external_key_lookup` mapping table to support both `uuid` and `int` foreign keys.
-- Fixed a bug in the `crmTrait` file where the Model instance was not given but the Model name (string).
+- Fixed a bug in the `CrmTrait` file where the Model instance was not given but the Model name (string).
 - Fixed PHP class types.
 - Documentation updates.